Use a less brittle but more duplicative approach for Location component
